pyproject.toml
[tool.poetry]
name = "py-xtdb"
version = "0.1.0"
description = ""
authors = ["joefromct <joefromct@fastmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.27.1"
cytoolz = "^0.11.2"
edn-format = "^0.7.5"
pampy = "^0.3.0"
pycurl = "^7.45.1"
faker-credit-score = "^0.3.0"
rootpath = "^0.1.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
ipython = "^8.3.0"
Faker = "^13.6.0"
tabulate = "^0.8.9"
hierarchy-utils = "^0.1.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
